+++ support/wsl-relay/relay-test.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+set -e
+
+RELAY_SOCKET=$(docker context inspect --format json | jq -e ".[0].Endpoints.docker.Host | sub(\"unix://\"; \"\")" | tr -d '"')
+
+echo "Building wsl-relay"
+go build --ldflags '-s -w -linkmode external -extldflags "-fno-PIC -static"' -buildmode pie -tags "osusergo netgo static_build"
+
+echo "Starting relay - listen to $RELAY_SOCKET and relay to $RELAY_PIPE"
+./wsl-relay --socket "$RELAY_SOCKET"